@@ -1,4 +1,4 @@
-import Vue from 'vue';
+import { createApp, h } from 'vue';
 import App from './App.vue';
 import './registerServiceWorker';
 import router from './router';
@@ -7,11 +7,7 @@ import vuetify from './plugins/vuetify';
 import 'roboto-fontface/css/roboto/roboto-fontface.css';
 import '@mdi/font/css/materialdesignicons.css';
 
-Vue.config.productionTip = false;
-
-new Vue({
-  router,
-  store,
+createApp({
   vuetify,
-  render: (h) => h(App),
-}).$mount('#app');
+  render: () => h(App),
+}).use(router).use(store).mount('#app');
